requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
geographiclib==2.1
geopy==2.4.1
sqlparse==0.5.3
tomli==2.3.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
